Move the servlets into an internal package.
